requirements.txt
streamlit
pandas
duckdb
groq
plotly
python-dotenv
langchain-huggingface
langchain-chroma